+++ Core/MigrationService.php
@@ -249,13 +249,15 @@ class MigrationService implements ContextProviderInterface
      * @param string $defaultLanguageCode
      * @param string|int|false|null $adminLogin when false, current user is used; when null, hardcoded admin account
      * @param bool $force when true, execute a migration if it was already in status DONE or SKIPPED (would throw by default)
+     * @param bool|null $forcesigchildHandling
      * @throws \Exception
      *
      * @todo treating a null and false $adminLogin values differently is prone to hard-to-track errors.
      *       Shall we use instead -1 to indicate the desire to not-login-as-admin-user-at-all ?
+     * @todo refactor there start to be too many parameters here. Move to an single parameter: array of options or value-object
      */
     public function executeMigration(MigrationDefinition $migrationDefinition, $useTransaction = true,
-        $defaultLanguageCode = null, $adminLogin = null, $force = false)
+        $defaultLanguageCode = null, $adminLogin = null, $force = false, $forceSigchildHandling = null)
     {
         if ($migrationDefinition->status == MigrationDefinition::STATUS_TO_PARSE) {
             $migrationDefinition = $this->parseMigrationDefinition($migrationDefinition);
@@ -266,7 +268,7 @@ class MigrationService implements ContextProviderInterface
         }
 
         /// @todo add support for setting in $migrationContext a userContentType ?
-        $migrationContext = $this->migrationContextFromParameters($defaultLanguageCode, $adminLogin);
+        $migrationContext = $this->migrationContextFromParameters($defaultLanguageCode, $adminLogin, $forceSigchildHandling);
 
         // set migration as begun - has to be in own db transaction
         $migration = $this->storageHandler->startMigration($migrationDefinition, $force);
@@ -473,9 +475,10 @@ class MigrationService implements ContextProviderInterface
     /**
      * @param string $defaultLanguageCode
      * @param string|int|false $adminLogin
-     * @return array
+     * @param bool|null
+     * @return array $forceSigchildHandling
      */
-    protected function migrationContextFromParameters($defaultLanguageCode = null, $adminLogin = null)
+    protected function migrationContextFromParameters($defaultLanguageCode = null, $adminLogin = null, $forceSigchildHandling = null )
     {
         $properties = array();
 
@@ -485,6 +488,10 @@ class MigrationService implements ContextProviderInterface
         // nb: other parts of the codebase treat differently a false and null values for $properties['adminUserLogin']
         if ($adminLogin !== null) {
             $properties['adminUserLogin'] = $adminLogin;
+        }
+        if ($forceSigchildHandling !== null)
+        {
+            $properties['forceSigchildHandling'] = $forceSigchildHandling;
         }
 
         return $properties;